Split server into http and ws
